# api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Blog System API
  description: |
    个人博客系统的RESTful API规范，支持Markdown文章管理、图片处理和文件操作。
    
    ## 功能特性
    - 文章CRUD操作
    - Markdown文件存储和检索
    - 图片上传和Base64转换
    - 文章搜索和筛选
    - 标签管理
    
  version: 1.0.0
  contact:
    name: Blog System API Support
    email: support@blog-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api.blog-system.com/v1
    description: 生产环境

tags:
  - name: articles
    description: 文章管理相关接口
  - name: images
    description: 图片处理相关接口
  - name: files
    description: 文件操作相关接口

paths:
  /articles:
    get:
      tags:
        - articles
      summary: 获取文章列表
      description: 支持分页、搜索和筛选的文章列表查询
      parameters:
        - name: page
          in: query
          description: 页码，从0开始
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: 每页数量
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: title
          in: query
          description: 按标题搜索
          required: false
          schema:
            type: string
        - name: tags
          in: query
          description: 按标签筛选，多个标签用逗号分隔
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: 按状态筛选
          required: false
          schema:
            $ref: '#/components/schemas/ArticleStatus'
        - name: sort
          in: query
          description: 排序字段
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt, publishedAt, title]
            default: createdAt
        - name: direction
          in: query
          description: 排序方向
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功获取文章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - articles
      summary: 创建新文章
      description: 创建一篇新的博客文章
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleRequest'
      responses:
        '201':
          description: 文章创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /articles/{id}:
    get:
      tags:
        - articles
      summary: 获取文章详情
      description: 根据ID获取文章的详细信息
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功获取文章详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - articles
      summary: 更新文章
      description: 更新指定ID的文章信息
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
      responses:
        '200':
          description: 文章更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - articles
      summary: 删除文章
      description: 删除指定ID的文章
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 文章删除成功
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /images:
    get:
      tags:
        - images
      summary: 获取图片列表
      description: 获取图片列表，支持分页
      parameters:
        - name: page
          in: query
          description: 页码，从0开始
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: 每页数量
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 成功获取图片列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - images
      summary: 上传图片
      description: 上传图片文件并转换为Base64格式存储
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片文件
                description:
                  type: string
                  description: 图片描述
              required:
                - file
      responses:
        '201':
          description: 图片上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: 文件过大
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /images/{id}:
    get:
      tags:
        - images
      summary: 获取图片信息
      description: 根据ID获取图片的详细信息和Base64内容
      parameters:
        - name: id
          in: path
          required: true
          description: 图片ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功获取图片信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - images
      summary: 删除图片
      description: 删除指定ID的图片
      parameters:
        - name: id
          in: path
          required: true
          description: 图片ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 图片删除成功
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /files/markdown/{articleId}:
    get:
      tags:
        - files
      summary: 获取Markdown文件内容
      description: 获取指定文章的Markdown源文件内容
      parameters:
        - name: articleId
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功获取Markdown内容
          content:
            text/plain:
              schema:
                type: string
                description: Markdown文件内容
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - files
      summary: 保存Markdown文件内容
      description: 保存文章的Markdown源文件内容
      parameters:
        - name: articleId
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: Markdown文件内容
      responses:
        '200':
          description: Markdown文件保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    ArticleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 文章唯一标识
        title:
          type: string
          description: 文章标题
        slug:
          type: string
          description: 文章URL友好标识
        summary:
          type: string
          description: 文章摘要
        contentPath:
          type: string
          description: Markdown文件存储路径
        status:
          $ref: '#/components/schemas/ArticleStatus'
        tags:
          type: array
          items:
            type: string
          description: 文章标签列表
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        publishedAt:
          type: string
          format: date-time
          description: 发布时间
          nullable: true
      required:
        - id
        - title
        - slug
        - contentPath
        - status
        - createdAt
        - updatedAt

    CreateArticleRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: 文章标题
        summary:
          type: string
          maxLength: 1000
          description: 文章摘要
        content:
          type: string
          description: Markdown内容
        tags:
          type: array
          items:
            type: string
          description: 文章标签列表
        status:
          $ref: '#/components/schemas/ArticleStatus'
      required:
        - title
        - content

    UpdateArticleRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: 文章标题
        summary:
          type: string
          maxLength: 1000
          description: 文章摘要
        content:
          type: string
          description: Markdown内容
        tags:
          type: array
          items:
            type: string
          description: 文章标签列表
        status:
          $ref: '#/components/schemas/ArticleStatus'

    ArticleStatus:
      type: string
      enum:
        - DRAFT
        - PUBLISHED
        - ARCHIVED
      description: |
        文章状态:
        * DRAFT - 草稿
        * PUBLISHED - 已发布
        * ARCHIVED - 已归档

    ArticlePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ArticleResponse'
          description: 文章列表
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
        totalElements:
          type: integer
          format: int64
          description: 总记录数
        totalPages:
          type: integer
          description: 总页数
        first:
          type: boolean
          description: 是否为第一页
        last:
          type: boolean
          description: 是否为最后一页
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last

    ImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 图片唯一标识
        filename:
          type: string
          description: 存储文件名
        originalName:
          type: string
          description: 原始文件名
        filePath:
          type: string
          description: 文件存储路径
        fileSize:
          type: integer
          format: int64
          description: 文件大小（字节）
        mimeType:
          type: string
          description: MIME类型
        base64Content:
          type: string
          description: Base64编码的图片内容
        createdAt:
          type: string
          format: date-time
          description: 上传时间
      required:
        - id
        - filename
        - originalName
        - filePath
        - fileSize
        - mimeType
        - base64Content
        - createdAt

    ImagePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ImageResponse'
          description: 图片列表
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
        totalElements:
          type: integer
          format: int64
          description: 总记录数
        totalPages:
          type: integer
          description: 总页数
        first:
          type: boolean
          description: 是否为第一页
        last:
          type: boolean
          description: 是否为最后一页

    FileOperationResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
        message:
          type: string
          description: 操作结果消息
        filePath:
          type: string
          description: 文件路径
        timestamp:
          type: string
          format: date-time
          description: 操作时间
      required:
        - success
        - message
        - timestamp

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: 错误代码
            message:
              type: string
              description: 错误消息
            details:
              type: string
              description: 错误详情
            timestamp:
              type: string
              format: date-time
              description: 错误发生时间
            path:
              type: string
              description: 请求路径
          required:
            - code
            - message
            - timestamp
      required:
        - error

  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "请求参数验证失败"
              details: "标题不能为空"
              timestamp: "2024-01-01T10:00:00Z"
              path: "/api/v1/articles"

    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "ARTICLE_NOT_FOUND"
              message: "文章不存在"
              details: "ID为123的文章未找到"
              timestamp: "2024-01-01T10:00:00Z"
              path: "/api/v1/articles/123"

    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_SERVER_ERROR"
              message: "服务器内部错误"
              details: "数据库连接失败"
              timestamp: "2024-01-01T10:00:00Z"
              path: "/api/v1/articles"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌

security:
  - BearerAuth: []